services/analytics-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.7
pymongo==4.5.0
redis==5.0.1
pandas==2.1.0
numpy==1.26.0
scipy==1.11.2
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.16.1
dash==2.13.0
statsmodels==0.14.0
python-jose==3.3.0
pyjwt==2.8.0
httpx==0.24.1
python-multipart==0.0.6
python-dotenv==1.0.0
apscheduler==3.10.1
pytz==2023.3
pytest==7.4.2
pytest-asyncio==0.21.1
aiofiles==23.2.1